# whisper.android: How to build with CLBlast