--- a/examples/whisper.android/lib/build.gradle
+++ b/examples/whisper.android/lib/build.gradle
@@ -16,6 +16,12 @@ android {
         ndk {
             abiFilters 'arm64-v8a', 'armeabi-v7a', 'x86', 'x86_64'
         }
+        externalNativeBuild {
+            cmake {
+                // Uncomment to use CLBlast on supported architectures.
+                // arguments '-DWHISPER_CLBLAST=ON'
+            }
+        }
     }
 
     buildTypes {