--- a/examples/whisper.android/lib/build.gradle
+++ b/examples/whisper.android/lib/build.gradle
@@ -18,18 +18,24 @@ android {
         }
         externalNativeBuild {
             cmake {
-                // When set, builds whisper.android against the version located at GGML_HOME
-                // instead of the version bundled with whisper.cpp.
-                //
-                // arguments "-DGGML_HOME=${rootDir.absolutePath}/../../../ggml",
+                // When set, builds whisper.android against the version located
+                // at GGML_HOME instead of the copy bundled with whisper.cpp.
+                if (
+                    project.hasProperty('GGML_HOME') &&
+                    project.findProperty('GGML_CLBLAST') == 'ON'
+                ) {
+                    // Turning on CLBlast requires GGML_HOME
+                    arguments "-DGGML_HOME=${project.property('GGML_HOME')}",
+                         "-DGGML_CLBLAST=ON",
+                         "-DOPENCL_LIB=${project.property('OPENCL_LIB')}",
+                         "-DCLBLAST_HOME=${project.property('CLBLAST_HOME')}",
+                         "-DOPENCL_ROOT=${project.property('OPENCL_ROOT')}",
+                         "-DCMAKE_FIND_ROOT_PATH_MODE_INCLUDE=BOTH",
+                         "-DCMAKE_FIND_ROOT_PATH_MODE_LIBRARY=BOTH"
+                } else if (project.hasProperty('GGML_HOME')) {
+                    arguments "-DGGML_HOME=${project.property('GGML_HOME')}"
+                }
 
-                // Turn on CLBlast on supported architectures (requires GGML_HOME):
-                //         "-DGGML_CLBLAST=ON"
-                //         "-DOPENCL_LIB=${rootDir.absolutePath}/lib/src/main/jniLibs/arm64-v8a/libOpenCL.so",
-                //         "-DCLBLAST_HOME=${rootDir.absolutePath}/../../../CLBlast",
-                //         "-DOPENCL_ROOT=${rootDir.absolutePath}/../../../OpenCL-Headers",
-                //         "-DCMAKE_FIND_ROOT_PATH_MODE_INCLUDE=BOTH",
-                //         "-DCMAKE_FIND_ROOT_PATH_MODE_LIBRARY=BOTH"
             }
         }
     }