--- a/examples/whisper.android/lib/src/main/jni/whisper/CMakeLists.txt
+++ b/examples/whisper.android/lib/src/main/jni/whisper/CMakeLists.txt
@@ -5,27 +5,28 @@ project(whisper.cpp)
 set(CMAKE_CXX_STANDARD 11)
 set(WHISPER_LIB_DIR ${CMAKE_SOURCE_DIR}/../../../../../../..)
 
+# Path to external GGML, otherwise uses the copy in whisper.cpp.
+option(GGML_HOME       "whisper: Path to external GGML source" OFF)
+
 set(
         SOURCE_FILES
         ${WHISPER_LIB_DIR}/whisper.cpp
         ${CMAKE_SOURCE_DIR}/jni.c
 )
 
-option(
-        GGML_HOME
-        "Path to GGML"
-        ${WHISPER_LIB_DIR}/../ggml
-)
-
-# Building from Android doesn't seem to respect the default value, so we set it here explicitly.
 if(NOT GGML_HOME)
-    set(GGML_HOME ${WHISPER_LIB_DIR}/../ggml)
+    set(
+        SOURCE_FILES
+        ${SOURCE_FILES}
+        ${WHISPER_LIB_DIR}/ggml.c
+        ${WHISPER_LIB_DIR}/ggml-alloc.c
+        ${WHISPER_LIB_DIR}/ggml-backend.c
+        ${WHISPER_LIB_DIR}/ggml-quants.c
+
+    )
 endif()
 
 find_library(LOG_LIB log)
-
-include(FetchContent)
-FetchContent_Declare(ggml SOURCE_DIR ${GGML_HOME})
 
 function(build_library target_name)
     add_library(
@@ -34,12 +35,12 @@ function(build_library target_name)
         ${SOURCE_FILES}
     )
 
-    target_link_libraries(${target_name} ${LOG_LIB} android ggml)
-
     if (${target_name} STREQUAL "whisper_v8fp16_va")
         target_compile_options(${target_name} PRIVATE -march=armv8.2-a+fp16)
+        set(GGML_COMPILE_OPTIONS                      -march=armv8.2-a+fp16)
     elseif (${target_name} STREQUAL "whisper_vfpv4")
         target_compile_options(${target_name} PRIVATE -mfpu=neon-vfpv4)
+        set(GGML_COMPILE_OPTIONS                      -mfpu=neon-vfpv4)
     endif ()
 
     if (NOT ${CMAKE_BUILD_TYPE} STREQUAL "Debug")
@@ -53,28 +54,27 @@ function(build_library target_name)
         target_link_options(${target_name} PRIVATE -flto)
 
     endif ()
+
+    if(GGML_HOME)
+        include(FetchContent)
+        FetchContent_Declare(ggml SOURCE_DIR ${GGML_HOME})
+        FetchContent_MakeAvailable(ggml)
+
+        target_compile_options(ggml PRIVATE ${GGML_COMPILE_OPTIONS})
+        target_link_libraries(${target_name} ${LOG_LIB} android ggml)
+    else()
+        target_link_libraries(${target_name} ${LOG_LIB} android)
+    endif()
+
+
 endfunction()
 
 if (${ANDROID_ABI} STREQUAL "arm64-v8a")
-    if (WHISPER_CLBLAST)
-        message(STATUS "Building with CLBlast")
-        set(CMAKE_FIND_ROOT_PATH_MODE_INCLUDE BOTH)
-        set(CMAKE_FIND_ROOT_PATH_MODE_LIBRARY BOTH)
-
-        set(GGML_CLBLAST ON CACHE INTERNAL "")
-        set(CLBLAST_HOME ${WHISPER_LIB_DIR}/../CLBlast)
-        set(OPENCL_ROOT ${WHISPER_LIB_DIR}/../OpenCL-Headers)
-        set(OPENCL_LIB ${CMAKE_SOURCE_DIR}/../../jniLibs/arm64-v8a/libOpenCL.so)
-    endif()
-
-    FetchContent_MakeAvailable(ggml)
     build_library("whisper_v8fp16_va")
 elseif (${ANDROID_ABI} STREQUAL "armeabi-v7a")
-    FetchContent_MakeAvailable(ggml)
     build_library("whisper_vfpv4")
-else()
-    FetchContent_MakeAvailable(ggml)
-    build_library("whisper") # Default target
 endif ()
+
+build_library("whisper") # Default target
 
 include_directories(${WHISPER_LIB_DIR})