Documenting for anyone else who wants to get CLBlast running on Android.

Depends on ggerganov/ggml#706.

Benchmark and transcription measurements below are from a Release variant built against ggerganov/ggml@53558f9cb9731e57134ebc86275ddf1eecf18356 (with CLBlast) and ggerganov/whisper.cpp@e72e4158debb04126a0fabedf0452a5551780ea0 (without CLBlast).

---

Benchmarks

Without CLBlast (`BLAS = 0`).

```
System Info: AVX = 0 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | METAL = 0 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | CUDA = 0 | COREML = 0 | OPENVINO = 0 | 
Loading data...
Copying jfk.wav...
All data copied to working directory.
Loading model...
Loaded model ggml-tiny.en-q5_1.bin.
Running benchmark. This will take minutes...
[...snip...]
  64 x   64: Q4_0     0.7 GFLOPS (128 runs) | Q4_1     0.6 GFLOPS (128 runs)
  64 x   64: Q5_0     0.6 GFLOPS (128 runs) | Q5_1     0.9 GFLOPS (128 runs) | Q8_0     1.2 GFLOPS (128 runs)
  64 x   64: F16      1.2 GFLOPS (128 runs) | F32      0.5 GFLOPS (128 runs)
 128 x  128: Q4_0     3.4 GFLOPS (128 runs) | Q4_1     4.2 GFLOPS (128 runs)
 128 x  128: Q5_0     5.2 GFLOPS (128 runs) | Q5_1     2.7 GFLOPS (128 runs) | Q8_0     5.6 GFLOPS (128 runs)
 128 x  128: F16      3.6 GFLOPS (128 runs) | F32      4.8 GFLOPS (128 runs)
 256 x  256: Q4_0    12.9 GFLOPS (128 runs) | Q4_1    12.3 GFLOPS (128 runs)
 256 x  256: Q5_0    11.2 GFLOPS (128 runs) | Q5_1    10.4 GFLOPS (128 runs) | Q8_0    13.0 GFLOPS (128 runs)
 256 x  256: F16     15.2 GFLOPS (128 runs) | F32      9.0 GFLOPS (128 runs)
 512 x  512: Q4_0    15.2 GFLOPS ( 57 runs) | Q4_1    16.6 GFLOPS ( 62 runs)
 512 x  512: Q5_0    12.6 GFLOPS ( 48 runs) | Q5_1    12.6 GFLOPS ( 47 runs) | Q8_0    16.1 GFLOPS ( 60 runs)
 512 x  512: F16     19.4 GFLOPS ( 73 runs) | F32     10.5 GFLOPS ( 40 runs)
1024 x 1024: Q4_0    19.1 GFLOPS (  9 runs) | Q4_1    20.2 GFLOPS ( 10 runs)
1024 x 1024: Q5_0    15.2 GFLOPS (  8 runs) | Q5_1    15.4 GFLOPS (  8 runs) | Q8_0    19.5 GFLOPS ( 10 runs)
1024 x 1024: F16     22.2 GFLOPS ( 11 runs) | F32     10.8 GFLOPS (  6 runs)
2048 x 2048: Q4_0    19.8 GFLOPS (  3 runs) | Q4_1    21.3 GFLOPS (  3 runs)
2048 x 2048: Q5_0    15.7 GFLOPS (  3 runs) | Q5_1    16.5 GFLOPS (  3 runs) | Q8_0    20.8 GFLOPS (  3 runs)
2048 x 2048: F16     23.3 GFLOPS (  3 runs) | F32     10.3 GFLOPS (  3 runs)
4096 x 4096: Q4_0    19.3 GFLOPS (  3 runs) | Q4_1    21.5 GFLOPS (  3 runs)
4096 x 4096: Q5_0    15.8 GFLOPS (  3 runs) | Q5_1    16.2 GFLOPS (  3 runs) | Q8_0    20.5 GFLOPS (  3 runs)
4096 x 4096: F16     20.2 GFLOPS (  3 runs) | F32      8.2 GFLOPS (  3 runs)
```

With CLBlast (`BLAS = 1`)

```
System Info: AVX = 0 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | METAL = 0 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | CUDA = 0 | COREML = 0 | OPENVINO = 0 | 
Loading data...
Copying jfk.wav...
All data copied to working directory.
Loading model...
Loaded model ggml-tiny.en-q5_1.bin.
Running benchmark. This will take minutes...
[...snip...]
  64 x   64: Q4_0     0.4 GFLOPS (128 runs) | Q4_1     0.4 GFLOPS (128 runs)
  64 x   64: Q5_0     0.3 GFLOPS (128 runs) | Q5_1     0.3 GFLOPS (128 runs) | Q8_0     0.4 GFLOPS (128 runs)
  64 x   64: F16      0.4 GFLOPS (128 runs) | F32      0.3 GFLOPS (128 runs)
 128 x  128: Q4_0     1.5 GFLOPS (128 runs) | Q4_1     1.8 GFLOPS (128 runs)
 128 x  128: Q5_0     2.1 GFLOPS (128 runs) | Q5_1     2.0 GFLOPS (128 runs) | Q8_0     2.0 GFLOPS (128 runs)
 128 x  128: F16      1.8 GFLOPS (128 runs) | F32      2.1 GFLOPS (128 runs)
 256 x  256: Q4_0     7.7 GFLOPS (128 runs) | Q4_1     7.9 GFLOPS (128 runs)
 256 x  256: Q5_0     7.7 GFLOPS (128 runs) | Q5_1     7.9 GFLOPS (128 runs) | Q8_0     7.6 GFLOPS (128 runs)
 256 x  256: F16      7.5 GFLOPS (128 runs) | F32      8.4 GFLOPS (128 runs)
 512 x  512: Q4_0    19.3 GFLOPS ( 73 runs) | Q4_1    18.8 GFLOPS ( 71 runs)
 512 x  512: Q5_0    19.3 GFLOPS ( 73 runs) | Q5_1    19.4 GFLOPS ( 73 runs) | Q8_0    19.4 GFLOPS ( 73 runs)
 512 x  512: F16     19.3 GFLOPS ( 73 runs) | F32     19.0 GFLOPS ( 71 runs)
1024 x 1024: Q4_0    63.0 GFLOPS ( 30 runs) | Q4_1    65.5 GFLOPS ( 31 runs)
1024 x 1024: Q5_0    63.3 GFLOPS ( 30 runs) | Q5_1    66.0 GFLOPS ( 31 runs) | Q8_0    66.5 GFLOPS ( 31 runs)
1024 x 1024: F16     63.0 GFLOPS ( 30 runs) | F32     62.2 GFLOPS ( 29 runs)
2048 x 2048: Q4_0   100.2 GFLOPS (  6 runs) | Q4_1    97.1 GFLOPS (  6 runs)
2048 x 2048: Q5_0   101.1 GFLOPS (  6 runs) | Q5_1    95.9 GFLOPS (  6 runs) | Q8_0   101.5 GFLOPS (  6 runs)
2048 x 2048: F16     96.7 GFLOPS (  6 runs) | F32    101.0 GFLOPS (  6 runs)
4096 x 4096: Q4_0   108.0 GFLOPS (  3 runs) | Q4_1    78.7 GFLOPS (  3 runs)
4096 x 4096: Q5_0    79.1 GFLOPS (  3 runs) | Q5_1    81.3 GFLOPS (  3 runs) | Q8_0    56.0 GFLOPS (  3 runs)
4096 x 4096: F16     64.9 GFLOPS (  3 runs) | F32     49.5 GFLOPS (  3 runs)
```

---

Transcribe `jfk.wav`.

`BLAS = 0`
```
whisper_print_timings:     load time =   183.27 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =    55.03 ms
whisper_print_timings:   sample time =    19.49 ms /     1 runs (   19.49 ms per run)
whisper_print_timings:   encode time =  1098.81 ms /     1 runs ( 1098.81 ms per run)
whisper_print_timings:   decode time =   290.64 ms /    27 runs (   10.76 ms per run)
whisper_print_timings:   batchd time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:   prompt time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:    total time =  1464.72 ms
```

`BLAS = 1`
```
whisper_print_timings:     load time =  2030.81 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =    49.70 ms
whisper_print_timings:   sample time =    26.18 ms /     1 runs (   26.18 ms per run)
whisper_print_timings:   encode time =  1344.75 ms /     1 runs ( 1344.75 ms per run)
whisper_print_timings:   decode time =   303.45 ms /    27 runs (   11.24 ms per run)
whisper_print_timings:   batchd time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:   prompt time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:    total time =  1725.54 ms
```