@bobqianic 2024-02-01
> Seems like quite a speed up, but notably does not translate to faster transcription.

Could you post the runtime information?

Example:

```
whisper_print_timings:     load time =  7241.31 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =    30.68 ms
whisper_print_timings:   sample time =   362.31 ms /   751 runs (    0.48 ms per run)
whisper_print_timings:   encode time =  1177.27 ms /     5 runs (  235.45 ms per run)
whisper_print_timings:   decode time =    79.06 ms /     5 runs (   15.81 ms per run)
whisper_print_timings:   batchd time =  3858.86 ms /   735 runs (    5.25 ms per run)
whisper_print_timings:   prompt time =   191.72 ms /    89 runs (    2.15 ms per run)
whisper_print_timings:    total time = 13115.86 ms
```