@gpokat 2024-02-02
did it get correct inference ? My build can not find anything when compile with CLBLAST on android.