@luciferous 2024-02-05
@bobqianic 

Looks like I was wrong about it not speeding up transcription 🙂.

The timing numbers below are from transcribing jfk.wav:

`BLAS = 0`
```
whisper_print_timings:     load time =   389.18 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   142.32 ms
whisper_print_timings:   sample time =    57.65 ms /     1 runs (   57.65 ms per run)
whisper_print_timings:   encode time = 28396.24 ms /     1 runs (28396.24 ms per run)
whisper_print_timings:   decode time =  2010.30 ms /    27 runs (   74.46 ms per run)
whisper_print_timings:   batchd time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:   prompt time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:    total time = 30621.52 ms
```

`BLAS = 1`
```
whisper_print_timings:     load time =   683.91 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   162.44 ms
whisper_print_timings:   sample time =    66.18 ms /     1 runs (   66.18 ms per run)
whisper_print_timings:   encode time =  3750.02 ms /     1 runs ( 3750.02 ms per run)
whisper_print_timings:   decode time =  2034.95 ms /    27 runs (   75.37 ms per run)
whisper_print_timings:   batchd time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:   prompt time =     0.00 ms /     1 runs (    0.00 ms per run)
whisper_print_timings:    total time =  6026.00 ms
```

I'll update the commit message to reflect.

Also to clarify my aims: since I don't understand the basis of the effect on performance, I'm hesitant to make any strong claims for my changes. My goal with this PR is only to merge a reference build to make it easier for further developments of CLBlast on Android.

<hr>

Android logs aren't hooked up to the transcription, so this is what I did to retrieve the `whisper_print_timings`.

```diff
diff --git a/examples/whisper.android/lib/src/main/jni/whisper/jni.c b/examples/whisper.android/lib/src/main/jni/whisper/jni.c
index 7f9d724..e522a3a 100644
--- a/examples/whisper.android/lib/src/main/jni/whisper/jni.c
+++ b/examples/whisper.android/lib/src/main/jni/whisper/jni.c
@@ -14,6 +14,13 @@
 #define LOGI(...) __android_log_print(ANDROID_LOG_INFO,     TAG, __VA_ARGS__)
 #define LOGW(...) __android_log_print(ANDROID_LOG_WARN,     TAG, __VA_ARGS__)
 
+static void log_callback(enum ggml_log_level level, const char * fmt, void * data) {
+    if (level == GGML_LOG_LEVEL_ERROR)     __android_log_print(ANDROID_LOG_ERROR, TAG, fmt, data);
+    else if (level == GGML_LOG_LEVEL_INFO) __android_log_print(ANDROID_LOG_INFO, TAG, fmt, data);
+    else if (level == GGML_LOG_LEVEL_WARN) __android_log_print(ANDROID_LOG_WARN, TAG, fmt, data);
+    else __android_log_print(ANDROID_LOG_DEFAULT, TAG, fmt, data);
+}
+
 static inline int min(int a, int b) {
     return (a < b) ? a : b;
 }
@@ -182,6 +189,8 @@ Java_com_whispercpp_whisper_WhisperLib_00024Companion_fullTranscribe(
     params.no_context = true;
     params.single_segment = false;
 
+    whisper_log_set(log_callback, NULL);
+
     whisper_reset_timings(context);
 
     LOGI("About to run whisper_full");
```