@luciferous 2024-02-05
@gpokat Transcribing `jfk.wav` seems to be working for me. What are you seeing?