@luciferous 2024-02-05
@gpokat Just want to verify that the `System Info` line show `BLAS = 1`, e.g.,
```
System Info: AVX = 0 | AVX2 = 0 | ... | BLAS = 1 | ...
```
If not, then, it might not have built correctly.