@luciferous 2024-02-05
@ggerganov I think these are in Debug. I'll follow up with numbers from a Release build.