@luciferous 2024-02-06
@ggerganov Updated the commit message with the new measurements. Curiously, changes to `ggml` in the last week or so may have degraded CLBlast benchmark performance. For the sake of posterity, this is what I saw building against a `ggml` version from sometime in mid-January (unfortunately, I don't recall which):

```
  64 x   64: Q4_0     0.0 GFLOPS ( 93 runs) | Q4_1     0.1 GFLOPS (107 runs)
  64 x   64: Q5_0     0.0 GFLOPS ( 89 runs) | Q5_1     0.1 GFLOPS ( 97 runs) | Q8_0     0.0 GFLOPS ( 87 runs)
  64 x   64: F16      0.1 GFLOPS (110 runs) | F32      0.0 GFLOPS ( 66 runs)
 128 x  128: Q4_0     0.4 GFLOPS ( 90 runs) | Q4_1     0.4 GFLOPS ( 91 runs)
 128 x  128: Q5_0     0.4 GFLOPS ( 89 runs) | Q5_1     0.4 GFLOPS ( 89 runs) | Q8_0     0.4 GFLOPS ( 91 runs)
 128 x  128: F16      0.4 GFLOPS (105 runs) | F32      0.4 GFLOPS ( 94 runs)
 256 x  256: Q4_0     2.3 GFLOPS ( 70 runs) | Q4_1     2.4 GFLOPS ( 73 runs)
 256 x  256: Q5_0     2.4 GFLOPS ( 71 runs) | Q5_1     2.4 GFLOPS ( 71 runs) | Q8_0     2.3 GFLOPS ( 70 runs)
 256 x  256: F16      2.3 GFLOPS ( 70 runs) | F32      2.6 GFLOPS ( 77 runs)
 512 x  512: Q4_0     9.6 GFLOPS ( 36 runs) | Q4_1     9.6 GFLOPS ( 36 runs)
 512 x  512: Q5_0     9.5 GFLOPS ( 36 runs) | Q5_1     9.7 GFLOPS ( 37 runs) | Q8_0     9.5 GFLOPS ( 36 runs)
 512 x  512: F16      6.5 GFLOPS ( 25 runs) | F32      9.6 GFLOPS ( 36 runs)
1024 x 1024: Q4_0    44.5 GFLOPS ( 21 runs) | Q4_1    42.8 GFLOPS ( 20 runs)
1024 x 1024: Q5_0    46.2 GFLOPS ( 22 runs) | Q5_1    45.7 GFLOPS ( 22 runs) | Q8_0    44.1 GFLOPS ( 21 runs)
1024 x 1024: F16     14.7 GFLOPS (  7 runs) | F32     36.0 GFLOPS ( 17 runs)
2048 x 2048: Q4_0    83.6 GFLOPS (  5 runs) | Q4_1    80.3 GFLOPS (  5 runs)
2048 x 2048: Q5_0    79.8 GFLOPS (  5 runs) | Q5_1    82.5 GFLOPS (  5 runs) | Q8_0    82.4 GFLOPS (  5 runs)
2048 x 2048: F16     53.8 GFLOPS (  4 runs) | F32     59.2 GFLOPS (  4 runs)
4096 x 4096: Q4_0   102.9 GFLOPS (  3 runs) | Q4_1   108.7 GFLOPS (  3 runs)
4096 x 4096: Q5_0   108.8 GFLOPS (  3 runs) | Q5_1   108.0 GFLOPS (  3 runs) | Q8_0   108.3 GFLOPS (  3 runs)
4096 x 4096: F16    183.6 GFLOPS (  3 runs) | F32    109.0 GFLOPS (  3 runs)
```