@gpokat 2024-02-06
@luciferous did you build CLBLAST with tuner or is your device already in a tuned list for clblast ? https://github.com/CNugteren/CLBlast/blob/master/doc/tuning.md
The untuned calculations could lead to opencl performance degradation.