@JunkFood02 2024-04-17
After compiling CLBlast with whisper.cpp and running it on my Snapdragon 8 Gen 2 device (with a tuned Adreno 740 GPU for CLBlast), benchmark results and transcription speed haven't shown significant changes (improvement or regression) compared to CPU inference without it. This could be because ggml isn't offloading many computations to the GPU/OpenCL. 

See also:
- https://github.com/ggerganov/llama.cpp/issues/5965
- https://github.com/ggerganov/llama.cpp/issues/5621
- https://github.com/ggerganov/llama.cpp/issues/6562